+++ CHANGES.md
@@ -1,5 +1,9 @@
 # Change Log
 
+#### 1.7.4 / 2017-10-25
+* add function and shortcode to check for a full group.
+* add new routines for adding and updating group admins through gravity forms.
+
 #### 1.7.3 / 2017-10-23
 * remove styles from plugin. Add to theme.